+++ src/modules/private/admin/applications/components/details/review/Overview.js
@@ -34,8 +34,8 @@ export default class Overview extends Component {
       OVERVIEW_FRM, formChangeWithIndex, confirmModal, toggleConfirmModal,
       removeData, confirmModalName,
     } = this.props.businessAppReviewStore;
-    const { myCapabilities } = this.props.userStore;
-    const isManager = myCapabilities.includes('APPLICATIONS_MANAGER');
+    const access = this.props.userStore.myAccessForModule('APPLICATIONS');
+    const isManager = access.asManager;
     const { businessApplicationDetailsAdmin } = this.props.businessAppStore;
     const { review } = businessApplicationDetailsAdmin;
     const submitted = (review && review.overview && review.overview.criticalPoint &&