+++ src/services/stores/queries/offerings/manage.js
@@ -1,8 +1,8 @@
 import gql from 'graphql-tag';
 
 export const allOfferingsCompact = gql`
-  query _getOfferings($stage: [OfferingStageEnumType]){
-    getOfferings(filters: { stage: $stage }){
+  query _getOfferings($stage: [OfferingStageEnumType], $issuerId: String!){
+    getOfferings(filters: { stage: $stage, issuerId: $issuerId }){
       id
       keyTerms {
         legalBusinessName
@@ -122,6 +122,10 @@ export const getOfferingDetails = gql`
           isPublic
         }
         logo {
+          url
+          isPublic
+        }
+        heroVideo {
           url
           isPublic
         }